// src/api/dto/category.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Genshin Quiz API
 * API for Genshin Impact Quiz Application
 * OpenAPI spec version: 1.0.0
 */

/**
 * 分类
 */
export type Category = (typeof Category)[keyof typeof Category];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const Category = {
  character: 'character',
  weapon: 'weapon',
  artifact: 'artifact',
  lore: 'lore',
  gameplay: 'gameplay',
  world: 'world',
  combat: 'combat',
  music: 'music',
  statistics: 'statistics',
  fun: 'fun',
  other: 'other',
} as const;