// src/api/dto/difficulty.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Genshin Quiz API
 * API for Genshin Impact Quiz Application
 * OpenAPI spec version: 1.0.0
 */

/**
 * 难度等级
 */
export type Difficulty = (typeof Difficulty)[keyof typeof Difficulty];

export const Difficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;